Fail Log July 30, Week4
This fail log is the same as I am posting on the blog. Actual commands and more detail are provided in the .md files I uploaded. I'm not sure if they are allowed to be the same, but I'm giving it a try this week. Doesn't make sense to write about the same thing twice.